Made Auto-renewal into a service so that the service can be restarted if the android os decides to kill it in the middle of auto-renewing
